+++ src/db_migrate.cpp
@@ -13,7 +13,7 @@
 namespace
 {
 
-constexpr auto version = std::string_view{"0.4.49"};
+constexpr auto version = std::string_view{"0.4.50"};
 
 auto print_help() -> void
 {